move shader data into separate module
