+++ crates/plinth/src/graphics/context.rs
@@ -17,8 +17,8 @@ use crate::graphics::TextureLoadError;
 use crate::graphics::draw::CanvasStorage;
 use crate::graphics::draw::DrawCommand;
 use crate::graphics::glyph_cache::GlyphCache;
-use crate::graphics::pipeline::DrawUniforms;
 use crate::graphics::pipeline::RenderPipelineCache;
+use crate::graphics::shader_data::DrawUniforms;
 use crate::graphics::surface::RenderError;
 use crate::graphics::surface::Surface;
 use crate::graphics::texture::StorageId;
