+++ crates/plinth/src/graphics/draw.rs
@@ -5,9 +5,9 @@ use crate::graphics::color::Color;
 use crate::graphics::glyph_cache::GlyphCache;
 use crate::graphics::paint::GradientPaint;
 use crate::graphics::paint::Paint;
-use crate::graphics::pipeline::GpuPaint;
-use crate::graphics::pipeline::GpuPrimitive;
-use crate::graphics::pipeline::PrimitiveRenderFlags;
+use crate::graphics::shader_data::GpuPaint;
+use crate::graphics::shader_data::GpuPrimitive;
+use crate::graphics::shader_data::PrimitiveRenderFlags;
 use crate::graphics::texture::StorageId;
 use crate::graphics::texture::Texture;
 use crate::graphics::texture::TextureManager;
